+++ src/main/java/com/example/swapit/config/AwsS3Config.java
@@ -61,18 +61,4 @@ public class AwsS3Config {
 			)) // IAM Role 또는 환경 변수 사용
 			.build();
 	}
-
-	@Bean
-	@Profile("test")
-	public S3Presigner testS3Presigner(
-		@Value("${cloud.aws.credentials.access-key}") String accessKey,
-		@Value("${cloud.aws.credentials.secret-key}") String secretKey
-	) {
-		return S3Presigner.builder()
-			.region(Region.AP_NORTHEAST_2) // S3 리전 설정 (서울)
-			.credentialsProvider(StaticCredentialsProvider.create(
-				AwsBasicCredentials.create(accessKey, secretKey)
-			)) // IAM Role 또는 환경 변수 사용
-			.build();
-	}
 }
