+++ src/test/java/com/example/swapit/domain/JpaAuditingTest.java
@@ -6,21 +6,15 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.test.context.ActiveProfiles;
 
-import com.example.swapit.config.TestFirebaseConfig;
-import com.example.swapit.config.TestS3Config;
+import com.example.swapit.config.TestConfig;
 import com.example.swapit.repository.CategoriesRepository;
 import com.example.swapit.repository.GoodsRepository;
 import com.example.swapit.repository.UsersRepository;
 import com.example.swapit.testcontainer.BaseIntegrationTest;
 
-import jakarta.transaction.Transactional;
-
 @SpringBootTest
-@Transactional
-@ActiveProfiles("test")
-@Import({TestFirebaseConfig.class, TestS3Config.class})
+@Import(TestConfig.class)
 class JpaAuditingTest extends BaseIntegrationTest {
 
 	@Autowired
